rename Tests Namespaces
