+++ composer.json
@@ -29,7 +29,7 @@
   },
   "autoload-dev": {
     "psr-4": {
-      "Tests\\": "tests/"
+      "Clicksports\\LexOffice\\Tests\\": "tests/"
     }
   },
   "minimum-stability": "dev",
