+++ tests/Voucher/ClientTest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Tests\Src\OrderConfirmation;
+namespace Clicksports\LexOffice\Tests\Voucher;
 
-use Clicksports\LexOffice\Exceptions\BadMethodCallException;
-use Clicksports\LexOffice\OrderConfirmation\Client;
+use Clicksports\LexOffice\Voucher\Client;
 use GuzzleHttp\Psr7\Response;
-use Tests\TestClient;
+use Clicksports\LexOffice\Tests\TestClient;
 
 class ClientTest extends TestClient
 {
@@ -53,46 +52,14 @@ class ClientTest extends TestClient
 
     public function testUpdate()
     {
-        $this->expectException(BadMethodCallException::class);
-
         $stub  = $this->createClientMockObject(
             Client::class,
             new Response(200, [], '{}'),
             ['update']
         );
 
-        $stub->update('resource-id', []);
-    }
+        $response = $stub->update('resource-id', []);
 
-    public function testDocument()
-    {
-        $stub  = $this->createClientMockObject(
-            Client::class,
-            new Response(200, [], '{"documentFileId": "fake-id"}'),
-            ['document']
-        );
-
-        $response = $stub->document('resource-id');
-
-        $this->assertEquals(
-            '{"documentFileId": "fake-id"}',
-            $response->getBody()->__toString()
-        );
-
-        $stub  = $this->createClientMultiMockObject(
-            Client::class,
-            [
-                new Response(200, [], '{"documentFileId": "fake-id"}'),
-                new Response(200, [], '{}')
-            ],
-            ['document']
-        );
-
-        $response = $stub->document('resource-id', true);
-
-        $this->assertEquals(
-            '{}',
-            $response->getBody()->__toString()
-        );
+        $this->assertEquals('{}', $response->getBody()->__toString());
     }
 }
